Add /etc/default/kdump-test-script handling
The environment variables LOCAL_ONLY and NO_RESULT are verified
from the running environment. This is fine when the script is run
interactively but when executed as an initscript, it is difficult
to preset these environment variable. Adding a file in /etc/default
to define those variable makes it easier.

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>